Fix snapshot serializer require, restructure pretty-format.
